Add customId for for animated movements and don't trigger movement when it isn't necessary
Adding an optional customId allows users to change the TileLayer's
tileUpdateTransformer behaviour based on what triggers an animated
movement.

In passing I noticed that it was possible to optimise the movement such
that no calls to the MapController are made if no movement occurs and if
both movement/rotation occur the MapControllers moveAndRotate method is
used which is more efficient than calling both move() and rotate().

I also added a check to make sure that if we animate movement that we
don't repeat the last movement (when animation.value == 1.0) and that we
use the AnimatedMoveId.finish id exactly once.